fix: 라운지 오브제 관련 수정
